# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Traffic Peak typing platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List all available lessons
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
      responses:
        "200":
          description: Lessons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson and its content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
  /lessons/{id}/progress:
    post:
      operationId: saveLessonProgress
      tags: [lessons]
      summary: Save user progress on a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonProgressInput"
      responses:
        "200":
          description: Progress saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"
...
  /hr/assessments:
    get:
      operationId: listHrAssessments
      tags: [hr]
      summary: List assessments created by the current user
      responses:
        "200":
          description: Assessment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HrAssessment"
        "401":
          description: Unauthorized
    post:
      operationId: createHrAssessment
      tags: [hr]
      summary: Create a new assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HrAssessmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrAssessment"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /hr/assessments/{id}:
    get:
      operationId: getHrAssessment
      tags: [hr]
      summary: Get assessment detail with results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrAssessmentDetail"
        "404":
          description: Not found
    put:
      operationId: updateHrAssessment
      tags: [hr]
      summary: Update assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HrAssessmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrAssessment"
    delete:
      operationId: deleteHrAssessment
      tags: [hr]
      summary: Delete assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /hr/assessments/{id}/results:
    get:
      operationId: listAssessmentResults
      tags: [hr]
      summary: Get candidate results for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Results list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateResult"
...
components:
  schemas:
...
    Lesson:
      type: object
      required: [id, title, level, category, content, order]
      properties:
        id:
          type: integer
        title:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        category:
          type: string
        content:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
    LessonProgressInput:
      type: object
      required: [wpm, accuracy, completed]
      properties:
        wpm:
          type: number
        accuracy:
          type: number
        completed:
          type: boolean
    LessonProgress:
      type: object
      required: [id, lessonId, wpm, accuracy, completed]
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        userId:
          type: ["integer", "null"]
        wpm:
          type: number
        accuracy:
          type: number
        completed:
          type: boolean
        bestWpm:
          type: number
        completedAt:
          type: ["string", "null"]

    # ── Leaderboard ───────────────────────────────────────────────────
...
    HrAssessmentInput:
      type: object
      required: [name, companyName, jobPosition]
      properties:
        name:
          type: string
        companyName:
          type: string
        jobPosition:
          type: string
        description:
          type: ["string", "null"]
        durationSeconds:
          type: integer
        difficulty:
          type: string
        language:
          type: string
        contentType:
          type: string
        customText:
          type: ["string", "null"]
        passingWpm:
          type: integer
        minAccuracy:
          type: number
        maxAttempts:
          type: integer
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
    HrAssessment:
      type: object
      required: [id, token, name, companyName, jobPosition, durationSeconds, difficulty, language, contentType, passingWpm, minAccuracy, maxAttempts, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
        token:
          type: string
        createdById:
          type: integer
        name:
          type: string
        companyName:
          type: string
        jobPosition:
          type: string
        description:
          type: ["string", "null"]
        durationSeconds:
          type: integer
        difficulty:
          type: string
        language:
          type: string
        contentType:
          type: string
        customText:
          type: ["string", "null"]
        passingWpm:
          type: integer
        minAccuracy:
          type: number
        maxAttempts:
          type: integer
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        candidateCount:
          type: integer
        passCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    CandidateResult:
      type: object
      required: [id, fullName, email, wpm, accuracy, passed, rank]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        tabSwitches:
          type: integer
        wpm:
          type: number
        cpm:
          type: number
        accuracy:
          type: number
        errorCount:
          type: integer
        passed:
          type: boolean
        rank:
          type: integer
        completedAt:
          type: ["string", "null"]
    HrAssessmentDetail:
      type: object
      required: [id, token, name, companyName, jobPosition, durationSeconds, passingWpm, minAccuracy, candidates]
      properties:
        id:
          type: integer
        token:
          type: string
        name:
          type: string
        companyName:
          type: string
        jobPosition:
          type: string
        description:
          type: ["string", "null"]
        durationSeconds:
          type: integer
        difficulty:
          type: string
        language:
          type: string
        contentType:
          type: string
        customText:
          type: ["string", "null"]
        passingWpm:
          type: integer
        minAccuracy:
          type: number
        maxAttempts:
          type: integer
        active:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        candidateCount:
          type: integer
        passCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/CandidateResult"
...